+++ src/component.spec.ts
@@ -1,10 +1,8 @@
-import { Component, run, interfaceOf, clone, ev } from './core'
+import { Component, run, clone } from './core'
 import {
   action,
   props,
-  vw,
   setGroup,
-  act,
   stateOf,
   spaceOf,
   sendMsg,
@@ -31,18 +29,6 @@ describe('Component helpers', () => {
 
     it('should accept an action-contextValue pair as first argument and a fetch value in the second', () => {
       expect(actionFn([['a1', 10], 7])).toEqual([10, 7])
-    })
-
-  })
-
-  describe('act function sugar for generic inputs', () => {
-    let ctx = {}
-    it('should return the same as ev without the input name', () => {
-      expect(act(<any> ctx, 'actionName', 's', 'value')).toEqual(ev(<any> ctx, 'action', ['actionName', 's'], 'value'))
-    })
-
-    it('should return the same as ev without the input name when context data is undefined', () => {
-      expect(act(<any> ctx, 'actionName', undefined, 'value')).toEqual(ev(<any> ctx, 'action', 'actionName', 'value'))
     })
 
   })
@@ -129,42 +115,6 @@ describe('Component helpers', () => {
       let newComp = setGroup('style', { base: { x: 2 } })(clone(comp))
       expect(newComp.groups.style.base.a).toBeUndefined()
       expect(newComp.groups.style.base.x).toEqual(2)
-    })
-
-  })
-
-  describe('vw function sugar for components', () => {
-    let child: Component<any> = {
-      name: 'Child',
-      state: {
-        count: 0,
-        data: 10,
-      },
-      inputs: ctx => ({}),
-      actions: {},
-      interfaces: {},
-    }
-    let comp: Component<any> = {
-      name: 'MyComp',
-      components: {
-        child,
-      },
-      state: {
-        count: 0,
-        data: 10,
-      },
-      inputs: ctx => ({}),
-      actions: {},
-      interfaces: {},
-    }
-    let app = run({
-      root: comp,
-      interfaces: {},
-    })
-
-    it ('should be the same to use vw and interfaceOf functions', () => {
-      let interfaceObj = vw(app.ctx.components['MyComp'].ctx, 'child')
-      expect(interfaceObj).toEqual(interfaceOf(app.ctx.components['MyComp'].ctx, 'child', 'view'))
     })
 
   })