+++ src/core/index.ts
@@ -1,3 +1,4 @@
 export * from './core'
 export * from './module'
+export * from './interface'
 export * from './handler'