+++ src/core/module.ts
@@ -4,16 +4,18 @@ import {
   Update,
   Context,
   ComponentSpaceIndex,
-  InputData,
   EventData,
   Executable,
   Components,
-  EventOptions,
+  Interfaces,
+  CtxInterfaceIndex,
 } from './core'
 import {
   HandlerInterface,
   HandlerObject,
 } from './handler'
+import { makeInterfaceHelpers, dispatch } from './interface'
+import { makeInputHelpers } from './inputs'
 
 export interface ModuleDef {
   root: Component<any>
@@ -103,21 +105,6 @@ export function createContext (ctx: Context, name: Identifier): Context {
   }
 }
 
-// gets an interface message from a certain component
-export function interfaceOf (ctx: Context, name: string, interfaceName: string): any {
-  let id = `${ctx.id}$${name}`
-  let componentSpace = ctx.components[id]
-  if (!componentSpace) {
-    ctx.error('interfaceOf', `there are no component space '${id}'`)
-    return {}
-  }
-  if (!componentSpace.def.interfaces[interfaceName]) {
-    ctx.error('interfaceOf', `there are no interface '${interfaceName}' in component '${componentSpace.def.name}' from space '${id}'`)
-    return {}
-  }
-  return componentSpace.def.interfaces[interfaceName](componentSpace.ctx, componentSpace.state)
-}
-
 // add a component to the component index
 export function nest (ctx: Context, name: Identifier, component: Component<any>, isStatic = false): Context {
   // create the global object for initialization
@@ -166,7 +153,8 @@ function _nest (ctx: Context, name: Identifier, component: Component<any>, isSta
     isStatic,
     // if state is an object, it is cloned
     state: typeof component.state === 'object' ? clone(component.state) : component.state,
-    inputs: component.inputs ? component.inputs(childCtx) : {},
+    inputs: component.inputs ? component.inputs(makeInputHelpers(childCtx)) : {},
+    interfaces: _makeInterfaces(childCtx, component.interfaces),
     components: clone(component.components || {}),
     def: component,
   }
@@ -182,6 +170,15 @@ function _nest (ctx: Context, name: Identifier, component: Component<any>, isSta
   }
 
   return childCtx
+}
+
+function _makeInterfaces (ctx: Context, interfaces: Interfaces): CtxInterfaceIndex {
+  let index: CtxInterfaceIndex = {}
+  let name
+  for (name in interfaces) {
+    index[name] = interfaces[name](makeInterfaceHelpers(ctx))
+  }
+  return index
 }
 
 function handleGroups (ctx: Context, component: Component<any>) {
@@ -243,99 +240,6 @@ export function unnestAll (ctx: Context, components: string[]) {
   for (let i = 0, len = components.length; i < len; i++) {
     unnest(ctx, components[i])
   }
-}
-
-// create an InputData array
-export function ev (
-  ctx: Context,
-  inputName: string,
-  context?: any,
-  param?: any,
-  options?: EventOptions
-): InputData {
-  return [ctx.id, inputName, context, param, options]
-}
-
-function computePath (path: any[], event) {
-  let data
-  let actual = event
-  for (let i = 0, len = path.length; i < len; i++) {
-    if (path[i] instanceof Array) {
-      data = {}
-      let keys = path[i]
-      for (let i = 0, len = keys.length; i < len; i++) {
-        data[keys[i]] = actual[keys[i]]
-      }
-    } else {
-      actual = actual[path[i]]
-    }
-  }
-  if (!data) {
-    data = actual
-  }
-  return data
-}
-
-export function computeEvent(event: any, iData: InputData): EventData {
-  let data
-
-  if (iData[3] === '*') {
-    // serialize the whole object (note that DOM events are not serializable, use paths instead)
-    data = JSON.parse(JSON.stringify(event))
-  } else if (event && iData[3] !== undefined) {
-    // have fetch parameter
-    if (iData[3] instanceof Array) {
-      // fetch parameter is a path, e.g. ['target', 'value']
-      let param = iData[3]
-      if (param[1] && param[1] instanceof Array) {
-        data = []
-        for (let i = 0, len = param.length; i < len; i++) {
-          data[i] = computePath(param[i], event)
-        }
-      } else {
-        // only one path
-        data = computePath(param, event)
-      }
-    } else {
-      // fetch parameter is only a getter, e.g. 'target'
-      data = event[iData[3]]
-    }
-  }
-  if (iData[2] === undefined && iData[3] === undefined) {
-    return [iData[0], iData[1]] // dispatch an input with no arguments
-  }
-  return [
-    iData[0], // component id
-    iData[1], // component event
-    iData[2], // context argument
-    data, // data
-    iData[2] !== undefined && iData[3] !== undefined
-      ? 'pair'
-      : (iData[2] !== undefined)
-      ? 'context'
-      : 'fn',
-  ]
-}
-
-// dispatch an input based on eventData to the respective component
-/* istanbul ignore next */
-// TODO: test propagation (CRITICAL)
-export const dispatch = (ctxIn: Context, eventData: EventData, isPropagated = true) => {
-  let id = eventData[0] + ''
-  // root component
-  let ctx = ctxIn.components[(id + '').split('$')[0]].ctx
-  let componentSpace = ctx.components[id]
-  if (!componentSpace) {
-    return ctx.error('dispatch', `there are no component space '${id}'`)
-  }
-  let inputName = eventData[1]
-  // extract data from eventData
-  let data = eventData[4] === 'pair' // is both?
-    ? [eventData[2], eventData[3]] // is both event data + context
-    : eventData[4] === 'context'
-    ? eventData[2] // is only context
-    : eventData[3] // is only event data
-  toIt(componentSpace.ctx, inputName, data)
 }
 
 export function propagate (ctx: Context, inputName: string, data: any) {
@@ -462,9 +366,9 @@ export function execute (ctx: Context, executable: void | Executable<any> | Exec
 export function notifyInterfaceHandlers (ctx: Context) {
   let space = ctx.components[ctx.id]
   let name
-  for (name in space.def.interfaces) {
+  for (name in space.interfaces) {
     if (ctx.interfaceHandlers[name]) {
-      ctx.interfaceHandlers[name].handle(space.def.interfaces[name](ctx, space.state))
+      ctx.interfaceHandlers[name].handle(space.interfaces[name](space.state))
     } else {
       // This only can happen when this method is called for a context that is not the root
       ctx.error('notifyInterfaceHandlers', `module does not have interface handler named '${name}' for component '${space.def.name}' from space '${ctx.id}'`)
@@ -574,6 +478,7 @@ export function run (moduleDef: ModuleDef): Module {
 
     // merges main component and ctx.id now contains it name
     ctx = nest(ctx, component.name, component, true)
+    let rootSpace = ctx.components[ctx.id]
 
     // middle function for hot-swapping
     if (middleFn) {
@@ -588,9 +493,9 @@ export function run (moduleDef: ModuleDef): Module {
 
     // pass initial value to each Interface Handler
     let name
-    for (name in component.interfaces) {
+    for (name in rootSpace.interfaces) {
       if (ctx.interfaceHandlers[name]) {
-        ctx.interfaceHandlers[name].handle(component.interfaces[name](ctx, ctx.components[rootName].state))
+        ctx.interfaceHandlers[name].handle(rootSpace.interfaces[name](ctx.components[rootName].state))
       } else {
         return ctx.error(
           'InterfaceHandlers',