+++ src/examples/complexList/Root.ts
@@ -1,5 +1,5 @@
-import { Actions, Inputs, ev, nest, unnest, clone, Interfaces, toIt } from '../../core'
-import { action, vw, props, toChild, stateOf } from '../../component'
+import { Actions, Inputs, nest, unnest, clone, Interfaces, toIt } from '../../core'
+import { action, props, toChild, stateOf } from '../../component'
 import { StyleGroup, clickable } from '../../style'
 import { assoc } from '../../fun'
 import { View, h } from '../../interfaces/view'
@@ -16,7 +16,7 @@ export const state = {
 
 export type S = typeof state
 
-export const inputs: Inputs<S> = ctx => ({
+export const inputs: Inputs<S> = ({ ctx }) => ({
   action: action(actions),
   inputKeyup: ([idx, [keyCode, text]]) => {
     if (keyCode === 13 && text !== '') {
@@ -64,7 +64,7 @@ export const actions: Actions<S> = {
   },
 }
 
-const view: View<S> = (ctx, s) => {
+const view: View<S> = ({ ctx, ev, vw }) => s => {
   let style = ctx.groups.style
 
   return h('div', {
@@ -76,7 +76,7 @@ const view: View<S> = (ctx, s) => {
       attrs: { placeholder: 'Type and hit enter' },
       props: { value: s.text },
       on: {
-        keyup: ev(ctx, 'inputKeyup', s.numItems, [
+        keyup: ev('inputKeyup', s.numItems, [
           ['keyCode'],
           ['target', 'value'],
         ]),
@@ -85,20 +85,20 @@ const view: View<S> = (ctx, s) => {
     h('div', {class: { [style.menuBar]: true }}, [
       h('div', {
         class: { [style.menuItem]: true },
-        on: { click: ev(ctx, 'setCheckAll', true) },
+        on: { click: ev('setCheckAll', true) },
       }, 'check all'),
       h('div', {
         class: { [style.menuItem]: true },
-        on: { click: ev(ctx, 'setCheckAll', false) },
+        on: { click: ev('setCheckAll', false) },
       }, 'uncheck all'),
       h('div', {
         class: { [style.menuItem]: true },
-        on: { click: ev(ctx, 'removeChecked') },
+        on: { click: ev('removeChecked') },
       }, 'remove checked'),
     ]),
     h('ul', {class: { [style.list]: true }},
       Object.keys(s.items).map(
-        idx => vw(ctx, idx),
+        idx => vw(idx),
       )
     ),
   ])